Support all versions of haplObs
Originally, the POM file only supported version 0.0.1-SNAPSHOT of haplObs.  But
since then, there have been a few new releases, with newer numbers.  And
version 0.0.1-SNAPSHOT is still being used for development builds.

Since the new versions of haplObs do not change the API, we can build against
any of those versions, so change the POM file to allow this.