Add text file explaining development release
This adds a new Markdown file.  The file first explains how to set up an
environment for doing development work.  In particular, it mentions the need
for haplObserve, which you either have to build yourself, or manually install.

Then, it explains how to configure Travis-CI for automated test and release.
You can either configure automated testing only, or you can configure both
automated test (when commits are pushed) and automated release (when a tag is
pushed).